+++ test/features/steps/iocommands.py
@@ -54,7 +54,6 @@ def step_tee_ouptut(context):
         os.remove(context.tee_file_name)
     context.cli.sendline('tee {0}'.format(
         os.path.basename(context.tee_file_name)))
-    wrappers.expect_pager(context, "\r\n", timeout=5)
 
 
 @when(u'we query "select 123456"')
@@ -66,14 +65,13 @@ def step_query_select_123456(context):
         +--------+\r
         | 123456 |\r
         +--------+\r
-        1 row in set\r
         """), timeout=5)
+    wrappers.expect_exact(context, '1 row in set', timeout=2)
 
 
 @when(u'we notee output')
 def step_notee_output(context):
     context.cli.sendline('notee')
-    wrappers.expect_pager(context, "\r\n", timeout=5)
 
 
 @then(u'we see 123456 in tee output')
@@ -82,4 +80,3 @@ def step_see_123456_in_ouput(context):
         assert '123456' in f.read()
     if os.path.exists(context.tee_file_name):
         os.remove(context.tee_file_name)
-    context.atprompt = True